% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_gbm_se.R
\name{build_gbm_se}
\alias{build_gbm_se}
\title{use curatedTCGAData and Brennan metadata to build an annotated GBM MAE}
\usage{
build_gbm_se(cache = BiocFileCache::BiocFileCache())
}
\arguments{
\item{cache}{if not NULL, location of BiocFileCache instance}
}
\description{
use curatedTCGAData and Brennan metadata to build an annotated GBM MAE
}
\note{
Will cache the created SummarizedExperiment instance unless told not to.
`rname` is `"brenn_gbm_se"`.
}
\examples{
gbmse = build_gbm_se()

# check effect of MGMT methylation on survival

xm = gbmse[, gbmse$mgmt_status !="" & gbmse$vital_status !=""]
library(survival)
ss = Surv(xm$os_days, 1*(xm$vital_status=="DECEASED"))
plot(survfit(ss~xm$mgmt_status), lty=1:2)
legend(1100, .95, lty=c(1,2), legend=c("MGMT methylated", "unmethylated"))
title("Time-on-study/vital status for 123 GBM patients analyzed in Brennan et al. PMID 24120142")
survdiff(ss~xm$mgmt_status)
}
