% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_brennan_clin.R
\name{get_brennan_clin}
\alias{get_brennan_clin}
\title{import clinical data from Brennan 2016}
\source{
\url{https://www.cell.com/cms/10.1016/j.cell.2013.09.034/attachment/9cefc2e8-caac-4225-bcdd-70f105ccf568/mmc7.xlsx}
}
\usage{
get_brennan_clin()
}
\description{
import clinical data from Brennan 2016
}
\note{
CSV extracted from XLSX; names normalized by janitor::clean_names, empty column deleted,
`case_id` added as rownames.
}
\examples{
head(get_brennan_clin)
}
